requirements.txt
numpy
tensorflow
keras
Pillow
opencv-python
streamlit